# Use inline assembly for md5