--- a/md5/src/lib.rs
+++ b/md5/src/lib.rs
@@ -12,17 +12,12 @@
 #[cfg(not(any(target_arch = "x86_64", target_arch = "x86")))]
 compile_error!("crate can only be used on x86 and x86-64 architectures");
 
-#[link(name = "md5", kind = "static")]
-extern "C" {
-    fn md5_compress(state: &mut [u32; 4], block: &[u8; 64]);
-}
+mod asm;
 
 /// Safe wrapper around assembly implementation of MD5 compression function
 #[inline]
 pub fn compress(state: &mut [u32; 4], blocks: &[[u8; 64]]) {
     for block in blocks {
-        unsafe {
-            md5_compress(state, block);
-        }
+        asm::md5_compress(state, block);
     }
 }