This is a md5 implementation using inline assembly instead of relying on `cc` and `.S`. This has the benefit of compiling and running on MSVC targets.

## Correctness
I have run a custom version of `md-5` pointing to this branch with `cargo +nightly test --features asm` and `cargo +nightly test --features asm --target i686-pc-windows-msvc` on a windows machine and passed the test.

## Performance
On my Linux machine, which has a Intel 6700K
Before:
```
test bench_compress ... bench:         104 ns/iter (+/- 0) = 9 MB/s
```

After:
```
test bench_compress ... bench:         103 ns/iter (+/- 0) = 9 MB/s
```

No observable performance degradation.

## Drawback
This change needs a MSRV bump to `1.59.0` where inline assembly was stabilized.