@johnmave126 2023-01-16
> @johnmave126 with inline assembly I think you can just add it upstream to the `md-5` crate

That makes sense. Forgot to check #45