@tarcieri 2023-01-16
Opened #55 to add info about that to the README